refactor(cu): parse raw tags into object before passing to process
I don't really like this, as this means that the CU has to
make decisions on how to parse the raw tags on the DataItem,
before evaluating the process.

I think it'd be better if the raw { name, value } tags were passed to WASM
then have utilities on the Lua side to extract the tag value based on tag name.
This would make it so the CU never parses tags, it simply extracts them,
then passes them into eval, leaving it up to the caller (Lua) to extract.

But the AOS process is expecting an object, which means the CU has to then
parse the tags into an object, so that's what i'm writing the CU to do.