feat(sdk)!: remove readState and implement readResult #148
BREAKING CHANGE: with the introduction of BiBo, readState is no longer
a useful api, since the state is an opaque byte array. Instead, in order
to "readState", a consumer should use sendMessage invoking a function
the process supports to obtain a view of the state in the message's eval
output. The consumer can then parse output according to its needs.